LayoutTests/platform/mac/fast/objc/dom-html-select-activate-expected.txt


This tests the Private Objective-C DOM HTML Bindings API for selecting indexes of a <select> that also activate the "change" event.

Select a single index in a normal <select>.
** change event fired **
PASS - expected index was selected.

Select a single index in a sized <select>.
** change event fired **
PASS - expected index was selected.

Select multiple indexes, in a <select multiple>.
** change event fired **
** change event fired **
PASS - selected options were what we expected.

Select multiple indexes, with multi-select on, in a <select multiple>.
** change event fired **
** change event fired **
PASS - selected options were what we expected.

Select multiple indexes, with multi-select on, in a <select multiple> with <optgroup>s.
** change event fired **
** change event fired **
PASS - selected options were what we expected.
